Service: Bugfix use before initialized panic (#1070)
## Description

Services are allowed to be const-initialized (but uninitialized) to
support stashing them in statics to be used in UEFI protocols. This is
typically safe as this static state won't begin being used until the
component starts and it can be initialized. In some rare cases (i.e. the
logger), we may attempt to use this static before the component has been
fully initialized. To resolve this, we add the map_or method, matching
the Option type's map_or interface, to allow the user to call the
service if it is initalized, otherwise use a default value.

This fixes an issue that stops all platforms using the advanced logger
from booting with no indication why, as we panic when attempting to use
the advanced logger, which is also how we log.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Booted to Shell, pass all CI tests

## Integration Instructions

N/A